+++ test/libsolidity/smtCheckerTests/invariants/loop_nested.sol
@@ -12,7 +12,8 @@ contract Simple {
 				++x;
 			assert(x == 10);
 		}
-		assert(y == x);
+		// Removed because of Spacer nondeterminism.
+		//assert(y == x);
 	}
 }
 // ----